@@ -1,15 +1,15 @@
 {
   "name": "transcriber-admin",
-  "version": "0.5.0",
+  "version": "0.5.2",
   "private": true,
   "dependencies": {
-    "@devexpress/dx-react-chart": "^2.0.4",
-    "@devexpress/dx-react-chart-material-ui": "^2.0.4",
-    "@devexpress/dx-react-core": "^2.0.4",
-    "@devexpress/dx-react-grid": "^2.0.4",
-    "@devexpress/dx-react-grid-material-ui": "^2.0.4",
-    "@material-ui/core": "^4.4.0",
-    "@material-ui/icons": "^4.2.1",
+    "@devexpress/dx-react-chart": "^2.2.1",
+    "@devexpress/dx-react-chart-material-ui": "^2.2.1",
+    "@devexpress/dx-react-core": "^2.2.1",
+    "@devexpress/dx-react-grid": "^2.2.1",
+    "@devexpress/dx-react-grid-material-ui": "^2.2.1",
+    "@material-ui/core": "^4.6.1",
+    "@material-ui/icons": "^4.5.1",
     "@orbit/coordinator": "^0.16.1",
     "@orbit/core": "^0.16.1",
     "@orbit/data": "^0.16.1",
@@ -21,37 +21,44 @@
     "aws-amplify": "^1.1.38",
     "axios": "^0.19.0",
     "js-base64": "^2.5.1",
+    "jwt-decode": "^2.2.0",
     "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "react": "^16.9.0",
-    "react-datasheet": "^1.3.14",
-    "react-dom": "^16.9.0",
+    "react": "^16.12.0",
+    "react-datasheet": "^1.4.0",
+    "react-dom": "^16.12.0",
     "react-file-drop": "^0.2.8",
     "react-localization": "^1.0.15",
     "react-orbitjs": "^0.1.15",
-    "react-redux": "^7.1.1",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "3.1.1",
-    "react-select": "^3.0.4",
+    "react-player": "^1.14.2",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "3.2.0",
+    "react-select": "^3.0.8",
     "react-string-replace": "^0.4.4",
-    "reactn": "^2.2.2",
+    "reactn": "^2.2.4",
     "redux": "^4.0.4",
     "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
     "reselect": "^4.0.0",
-    "typescript": "3.6.2"
+    "typescript": "3.7.2"
   },
   "scripts": {
+    "analyze": "source-map-explorer 'build/static/js/*.js'",
     "clean": "rimraf build converage",
     "start": "react-scripts start",
-    "build": "react-scripts build",
+    "build": "react-scripts --max_old_space_size=8192 build",
     "test": "react-scripts test",
     "deploy": "npm run stamp && amplify publish",
     "deploySetup": "amplify init && git checkout amplify && git checkout .gitignore",
     "eject": "react-scripts eject",
     "stamp": "node writeDate.js",
+    "appdev": "amplify env checkout appdev && node changeEnv.js appdev",
+    "appqa": "amplify env checkout appqa && node changeEnv.js appqa",
+    "appprod": "amplify env checkout appprod && node changeEnv.js appprod",
     "dev": "amplify env checkout dev && node changeEnv.js dev",
-    "qa": "amplify env checkout qa && node changeEnv.js qa"
+    "qa": "amplify env checkout qa && node changeEnv.js qa",
+    "prod": "amplify env checkout prod && node changeEnv.js prod"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -69,19 +76,21 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^4.1.0",
-    "@testing-library/react": "^9.1.4",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
     "@types/auth0-js": "^9.10.6",
-    "@types/jest": "24.0.18",
+    "@types/jest": "24.0.23",
     "@types/js-base64": "^2.3.1",
+    "@types/jwt-decode": "^2.2.1",
     "@types/lodash": "^4.14.138",
-    "@types/node": "12.7.4",
-    "@types/react": "16.9.2",
-    "@types/react-dom": "16.9.0",
-    "@types/react-redux": "^7.1.2",
-    "@types/react-router-dom": "^4.3.5",
-    "@types/react-select": "^3.0.4",
+    "@types/node": "12.12.11",
+    "@types/react": "16.9.11",
+    "@types/react-dom": "16.9.4",
+    "@types/react-redux": "^7.1.5",
+    "@types/react-router-dom": "^5.1.2",
+    "@types/react-select": "^3.0.8",
     "rimraf": "^3.0.0",
+    "source-map-explorer": "^2.1.1",
     "write": "^2.0.0"
   },
   "prettier": {
