@@ -4,14 +4,17 @@ import Access from './routes/Access';
 import Logout from './routes/Logout';
 import Drawer from './routes/drawer';
 import Loading from './routes/Loading';
+import Profile from './routes/Profile';
 import Callback from './callback/Callback';
+import TokenCheck from './hoc/TokenCheck';
 import Auth from './auth/Auth';
 import { parseQuery } from './utils/parseQuery';
 
 const auth = new Auth();
 
-const handleAuthentication = (props: any) => {
+const handleParameters = (props: any) => {
   const { location } = props;
+
   if (location.search !== '') {
     const params = parseQuery(location.search);
     if (params.inviteId && typeof params.inviteId === 'string') {
@@ -38,34 +41,43 @@ const handleNewOrg = (props: any) => {
 class App extends React.Component {
   render() {
     return (
-      <>
-        <Route
-          path="/"
-          exact={true}
-          render={props => <Access auth={auth} {...props} />}
-        />
-        <Route
-          path="/logout"
-          exact={true}
-          render={props => <Logout auth={auth} {...props} />}
-        />
-        <Route
-          path="/loading"
-          render={props => <Loading auth={auth} {...props} />}
-        />
-        <Route
-          path="/main"
-          render={props => <Drawer auth={auth} {...props} />}
-        />
-        <Route
-          path="/callback"
-          render={props => {
-            handleAuthentication(props);
-            return <Callback {...props} />;
-          }}
-        />
-        <Route path="/neworg" render={props => handleNewOrg(props)} />
-      </>
+      <TokenCheck auth={auth}>
+        <>
+          <Route
+            path="/"
+            exact={true}
+            render={props => {
+              handleParameters(props);
+              return <Access auth={auth} {...props} />;
+            }}
+          />
+          <Route
+            path="/logout"
+            exact={true}
+            render={props => <Logout auth={auth} {...props} />}
+          />
+          <Route
+            path="/loading"
+            render={props => <Loading auth={auth} {...props} />}
+          />
+          <Route
+            path="/main"
+            render={props => <Drawer auth={auth} {...props} />}
+          />
+          <Route
+            path="/profile"
+            render={props => <Profile auth={auth} {...props} />}
+          />
+          <Route
+            path="/callback"
+            render={props => {
+              handleParameters(props);
+              return <Callback {...props} />;
+            }}
+          />
+          <Route path="/neworg" render={props => handleNewOrg(props)} />
+        </>
+      </TokenCheck>
     );
   }
 }
