@@ -1,5 +1,5 @@
 import { Base64 } from 'js-base64';
-import { User } from './model';
+import { IApiError } from './model';
 import Coordinator, {
   RequestStrategy,
   SyncStrategy,
@@ -9,7 +9,7 @@ import Coordinator, {
 import IndexedDBSource from '@orbit/indexeddb';
 import IndexedDBBucket from '@orbit/indexeddb-bucket';
 import JSONAPISource from '@orbit/jsonapi';
-import { Schema, KeyMap, Transform } from '@orbit/data';
+import { Schema, KeyMap, Transform, NetworkError } from '@orbit/data';
 import { Bucket } from '@orbit/core';
 import Memory from '@orbit/memory';
 import Auth from './auth/Auth';
@@ -18,7 +18,7 @@ import { JSONAPISerializerCustom } from './serializers/JSONAPISerializerCustom';
 
 // import { Online } from './utils';
 
-const Sources = async (
+export const Sources = async (
   schema: Schema,
   memory: Memory,
   keyMap: KeyMap,
@@ -26,7 +26,10 @@ const Sources = async (
   setUser: (id: string) => void,
   setBucket: (bucket: Bucket) => void,
   setRemote: (remote: JSONAPISource) => void,
-  setCompleted: (valud: number) => void
+  setCompleted: (valud: number) => void,
+  InviteUser: (remote: JSONAPISource) => Promise<void>,
+  tableLoaded: (name: string) => void,
+  orbitError: (ex: IApiError) => void
 ) => {
   const tokenPart = auth.accessToken ? auth.accessToken.split('.') : [];
   const tokData = JSON.parse(
@@ -46,7 +49,7 @@ const Sources = async (
     namespace: 'transcriber',
   });
 
-  if (tokData.sub === userToken) {
+  if (tokData.sub === userToken || API_CONFIG.offline) {
     await backup
       .pull(q => q.findRecords())
       .then(transform => memory.sync(transform));
@@ -79,16 +82,16 @@ const Sources = async (
   }
 
   const coordinator = new Coordinator();
-  await coordinator.addSource(memory);
-  await coordinator.addSource(backup);
+  coordinator.addSource(memory);
+  coordinator.addSource(backup);
 
   if (!API_CONFIG.offline) {
-    await coordinator.addSource(remote);
+    coordinator.addSource(remote);
   }
 
   // Update indexedDb when memory updated
   // TODO: error if we can't read and write the indexedDB
-  await coordinator.addStrategy(
+  coordinator.addStrategy(
     new SyncStrategy({
       source: 'memory',
       target: 'backup',
@@ -97,8 +100,25 @@ const Sources = async (
   );
 
   if (!API_CONFIG.offline) {
+    // Trap error querying data (token expired or offline)
+    coordinator.addStrategy(
+      new RequestStrategy({
+        name: 'remote-pull-fail',
+
+        source: 'remote',
+        on: 'pullFail',
+
+        action(transform: Transform, ex: IApiError) {
+          console.log('***** api pull fail', transform, ex);
+          orbitError(ex);
+        },
+
+        blocking: true,
+      })
+    );
+
     // Query the remote server whenever the memory is queried
-    await coordinator.addStrategy(
+    coordinator.addStrategy(
       new RequestStrategy({
         name: 'remote-request',
 
@@ -112,8 +132,51 @@ const Sources = async (
       })
     );
 
+    // Trap error updating data (token expired or offline)
+    // See: https://github.com/orbitjs/todomvc-ember-orbit
+    coordinator.addStrategy(
+      new RequestStrategy({
+        name: 'remote-push-fail',
+
+        source: 'remote',
+        on: 'pushFail',
+
+        action(transform: Transform, ex: IApiError) {
+          const remote = coordinator.getSource('remote');
+          const memory = coordinator.getSource('memory') as Memory;
+
+          if (ex instanceof NetworkError) {
+            // When network errors are encountered, try again in 3s
+            console.log('NetworkError - will try again soon');
+            setTimeout(() => {
+              remote.requestQueue.retry();
+            }, 3000);
+          } else {
+            // When non-network errors occur, notify the user and
+            // reset state.
+            let label = transform.options && transform.options.label;
+            if (label) {
+              alert(`Unable to complete "${label}"`);
+            } else {
+              alert(`Unable to complete operation`);
+            }
+
+            // Roll back memory to position before transform
+            if (memory.transformLog.contains(transform.id)) {
+              console.log('Rolling back - transform:', transform.id);
+              memory.rollback(transform.id, -1);
+            }
+
+            return remote.requestQueue.skip();
+          }
+        },
+
+        blocking: true,
+      })
+    );
+
     // Update the remote server whenever the memory is updated
-    await coordinator.addStrategy(
+    coordinator.addStrategy(
       new RequestStrategy({
         name: 'remote-update',
 
@@ -128,7 +191,7 @@ const Sources = async (
     );
 
     // Sync all changes received from the remote server to the memory
-    await coordinator.addStrategy(
+    coordinator.addStrategy(
       new SyncStrategy({
         name: 'remote-sync',
 
@@ -140,25 +203,18 @@ const Sources = async (
     );
   }
 
-  await coordinator.addStrategy(new EventLoggingStrategy());
+  coordinator.addStrategy(new EventLoggingStrategy());
 
-  await coordinator.activate({ logLevel: LogLevel.Warnings }).then(() => {
+  coordinator.activate({ logLevel: LogLevel.Warnings }).then(() => {
     console.log('Coordinator will log warnings');
-    if (API_CONFIG.offline) {
-      memory
-        .query(q => q.findRecords('user'))
-        .then((u: Array<User>) => {
-          u = u.filter(u1 => u1.attributes && u1.attributes.name);
-          if (u.length === 1) {
-            setUser(u[0].id);
-          }
-        });
-    } else if (userToken === tokData.sub) {
+    if (userToken === tokData.sub || API_CONFIG.offline) {
       setUser(localStorage.getItem('user-id') as string);
     }
   });
 
   if (!API_CONFIG.offline && userToken !== tokData.sub) {
+    await InviteUser(remote);
+
     await remote
       .pull(q => q.findRecords('currentuser'))
       .then((transform: Transform[]) => {
@@ -174,80 +230,77 @@ const Sources = async (
     await remote
       .pull(q => q.findRecords('organization'))
       .then(transform => memory.sync(transform))
-      .then(() => setCompleted(10));
+      .then(() => setCompleted(14));
     await remote
       .pull(q => q.findRecords('organizationmembership'))
       .then(transform => memory.sync(transform))
-      .then(() => setCompleted(15));
-    await remote
-      .pull(q => q.findRecords('project'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(20));
-    await remote
-      .pull(q => q.findRecords('integration'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(25));
-    await remote
-      .pull(q => q.findRecords('invitation'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(30));
-    await remote
-      .pull(q => q.findRecords('projectintegration'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(35));
-    await remote
-      .pull(q => q.findRecords('projecttype'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(40));
-    await remote
-      .pull(q => q.findRecords('plan'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(45));
-    await remote
-      .pull(q => q.findRecords('plantype'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(50));
-    await remote
-      .pull(q => q.findRecords('section'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(55));
-    await remote
-      .pull(q => q.findRecords('passagesection'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(60));
-    await remote
-      .pull(q => q.findRecords('passage'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(65));
-    await remote
-      .pull(q => q.findRecords('userpassage'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(70));
+      .then(() => setCompleted(28));
     await remote
       .pull(q => q.findRecords('group'))
       .then(transform => memory.sync(transform))
-      .then(() => setCompleted(75));
+      .then(() => setCompleted(42));
     await remote
       .pull(q => q.findRecords('groupmembership'))
       .then(transform => memory.sync(transform))
-      .then(() => setCompleted(80));
+      .then(() => setCompleted(56));
+    await remote
+      .pull(q => q.findRecords('project'))
+      .then(transform => memory.sync(transform))
+      .then(() => setCompleted(70));
     await remote
       .pull(q => q.findRecords('role'))
       .then(transform => memory.sync(transform))
-      .then(() => setCompleted(85));
+      .then(() => setCompleted(84));
     await remote
-      .pull(q => q.findRecords('mediafile'))
-      .then(transform => memory.sync(transform))
-      .then(() => setCompleted(90));
-    await remote
-      .pull(q => q.findRecords('activitystate'))
+      .pull(q => q.findRecords('plan'))
       .then(transform => {
         memory.sync(transform);
         if (tokData.sub !== '') {
           localStorage.setItem('user-token', tokData.sub);
         }
+        tableLoaded('plan'); // If we are loading, length of tableLoad > 0
       })
       .then(() => setCompleted(95));
+    remote
+      .pull(q => q.findRecords('section'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('section'));
+    remote
+      .pull(q => q.findRecords('passagesection'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('passagesection'));
+    remote
+      .pull(q => q.findRecords('passage'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('passage'));
+    remote
+      .pull(q => q.findRecords('mediafile'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('mediafile'));
+    remote
+      .pull(q => q.findRecords('plantype'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('plantype'));
+    remote
+      .pull(q => q.findRecords('integration'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('integration'));
+    remote
+      .pull(q => q.findRecords('projectintegration'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('projectintegration'));
+    remote
+      .pull(q => q.findRecords('invitation'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('invitation'));
+    remote
+      .pull(q => q.findRecords('projecttype'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('projecttype'));
+    remote
+      .pull(q => q.findRecords('activitystate'))
+      .then(transform => memory.sync(transform))
+      .then(() => tableLoaded('activitystate'));
   }
 
   setCompleted(100);
