@@ -26,8 +26,12 @@ const memory = new Memory({ schema, keyMap });
 
 const globals = {
   organization: null,
+  orgRole: 'admin',
   project: null,
+  projRole: 'admin',
   plan: null,
+  tab: 0,
+  group: '',
   user: null,
   lang: 'en',
   memory: memory,
@@ -156,7 +160,6 @@ test('Clicking Add FAB on PlanTable displays dialogue', async () => {
   );
   await waitForElement(() => getByText(/^Genesis$/i));
   fireEvent.click(getByText('Add Plan'));
-  await waitForElement(() => getByText(/^Add a Plan$/i));
-
-  expect(getByText(/^Add a Plan$/i)).toHaveTextContent('Add a Plan');
+  await waitForElement(() => getByText(/^Type the name/i));
+  expect(getByText('Add')).toBeDisabled();
 });
