@@ -147,10 +147,14 @@ interface IStateProps {
 }
 interface IProps extends IStateProps {
   columns: Array<Column>;
-  columnWidths: Array<TableColumnWidthInfo>;
+  columnWidths?: Array<TableColumnWidthInfo>;
+  columnFormatting?: Table.ColumnExtension[];
   columnSorting?: Array<IntegratedSorting.ColumnExtension>;
+  sortingEnabled?: Array<SortingState.ColumnExtension>;
+  filteringEnabled?: Array<FilteringState.ColumnExtension>;
   defaultHiddenColumnNames?: Array<string>;
   dataCell?: any;
+  noDataCell?: any;
   numCols?: Array<string>;
   rows: Array<any>;
   sorting?: Array<Sorting>;
@@ -163,9 +167,13 @@ export function ShapingTable(props: IProps) {
     t,
     columns,
     columnWidths,
-    columnSorting,
+    columnFormatting,
+    columnSorting /* special sort function for each column as needed */,
+    sortingEnabled /* whether sorting is enabled for each column */,
+    filteringEnabled /* whether filtering is enabled for each column */,
     defaultHiddenColumnNames,
     dataCell,
+    noDataCell,
     numCols,
     rows,
     sorting,
@@ -183,9 +191,13 @@ export function ShapingTable(props: IProps) {
   return (
     <Grid rows={rows} columns={columns}>
       <FilteringState
-      // defaultFilters={[{ columnName: 'sectionId', operation: 'equal', value: '' }]}
+        columnExtensions={filteringEnabled ? filteringEnabled : []}
+        // defaultFilters={[{ columnName: 'sectionId', operation: 'equal', value: '' }]}
       />
-      <SortingState defaultSorting={sorting ? sorting : Array<Sorting>()} />
+      <SortingState
+        defaultSorting={sorting ? sorting : Array<Sorting>()}
+        columnExtensions={sortingEnabled ? sortingEnabled : []}
+      />
 
       <SelectionState onSelectionChange={handleSelect} />
 
@@ -208,16 +220,43 @@ export function ShapingTable(props: IProps) {
 
       <DragDropProvider />
 
-      {dataCell ? <Table cellComponent={dataCell} /> : <Table />}
+      {dataCell && noDataCell && !columnFormatting ? (
+        <Table cellComponent={dataCell} noDataCellComponent={noDataCell} />
+      ) : dataCell && !noDataCell && !columnFormatting ? (
+        <Table cellComponent={dataCell} />
+      ) : !dataCell && noDataCell && !columnFormatting ? (
+        <Table noDataCellComponent={noDataCell} />
+      ) : dataCell && noDataCell && columnFormatting ? (
+        <Table
+          cellComponent={dataCell}
+          noDataCellComponent={noDataCell}
+          columnExtensions={columnFormatting}
+        />
+      ) : dataCell && !noDataCell && columnFormatting ? (
+        <Table cellComponent={dataCell} columnExtensions={columnFormatting} />
+      ) : !dataCell && noDataCell && columnFormatting ? (
+        <Table
+          noDataCellComponent={noDataCell}
+          columnExtensions={columnFormatting}
+        />
+      ) : !dataCell && !noDataCell && columnFormatting ? (
+        <Table columnExtensions={columnFormatting} />
+      ) : (
+        <Table />
+      )}
       <TableColumnVisibility
-        defaultHiddenColumnNames={defaultHiddenColumnNames}
+        hiddenColumnNames={
+          defaultHiddenColumnNames ? defaultHiddenColumnNames : []
+        }
         emptyMessageComponent={noCols}
       />
-      <TableColumnResizing
-        minColumnWidth={50}
-        defaultColumnWidths={columnWidths}
-      />
-      <TableSelection showSelectAll={true} />
+      {columnWidths && (
+        <TableColumnResizing
+          minColumnWidth={50}
+          defaultColumnWidths={columnWidths}
+        />
+      )}
+      {!select || <TableSelection showSelectAll={true} />}
 
       <TableHeaderRow showSortingControls={true} />
       {shaping !== null && !shaping ? (
