@@ -25,7 +25,6 @@ import TreeGrid from './TreeGrid';
 import TranscriptionShow from './TranscriptionShow';
 import related from '../utils/related';
 import Auth from '../auth/Auth';
-import UserPassage from '../model/userPassage';
 import {
   sectionNumber,
   sectionReviewerName,
@@ -38,9 +37,6 @@ const useStyles = makeStyles((theme: Theme) =>
   createStyles({
     container: {
       display: 'flex',
-      marginLeft: theme.spacing(4),
-      marginRight: theme.spacing(4),
-      marginBottom: theme.spacing(4),
     },
     paper: {},
     actions: theme.mixins.gutters({
@@ -170,7 +166,6 @@ interface IStateProps {
 }
 
 interface IRecordProps {
-  userPassages: Array<UserPassage>;
   passages: Array<Passage>;
   passageSections: Array<PassageSection>;
   sections: Array<Section>;
@@ -190,7 +185,6 @@ export function TranscriptionTab(props: IProps) {
     passages,
     passageSections,
     sections,
-    userPassages,
     users,
     roles,
   } = props;
@@ -216,14 +210,12 @@ export function TranscriptionTab(props: IProps) {
   ];
 
   const [filter, setFilter] = useState(false);
-  const [group, setGroup] = useState(false);
 
   const handleMessageReset = () => {
     setMessage(<></>);
   };
 
   const handleFilter = () => setFilter(!filter);
-  const handleGroup = () => setGroup(!group);
 
   const handleSelect = (passageId: string) => () => {
     setPassageId(passageId);
@@ -243,16 +235,7 @@ export function TranscriptionTab(props: IProps) {
         activityState
       )
     );
-  }, [
-    plan,
-    userPassages,
-    passages,
-    passageSections,
-    sections,
-    users,
-    roles,
-    activityState,
-  ]);
+  }, [plan, passages, passageSections, sections, users, roles, activityState]);
 
   const LinkCell = ({ value, style, ...restProps }: any) => (
     <Table.Cell {...restProps} style={{ ...style }} value>
@@ -303,22 +286,6 @@ export function TranscriptionTab(props: IProps) {
               <FilterIcon className={classes.icon} />
             )}
           </Button>
-          <Button
-            key="group"
-            aria-label={t.group}
-            variant="outlined"
-            color="primary"
-            className={classes.button}
-            onClick={handleGroup}
-            title={'Show/Hide group panel'}
-          >
-            {t.group}
-            {group ? (
-              <SelectAllIcon className={classes.icon} />
-            ) : (
-              <FilterIcon className={classes.icon} />
-            )}
-          </Button>
         </div>
         <TreeGrid
           columns={columnDefs}
@@ -326,7 +293,7 @@ export function TranscriptionTab(props: IProps) {
           rows={data}
           getChildRows={getChildRows}
           cellComponent={Cell}
-          pageSizes={[5, 10, 20]}
+          pageSizes={[]}
           tableColumnExtensions={[
             { columnName: 'passages', align: 'right' },
             { columnName: 'name', wordWrapEnabled: true },
@@ -338,7 +305,7 @@ export function TranscriptionTab(props: IProps) {
           sorting={[{ columnName: 'name', direction: 'asc' }]}
           treeColumn={'name'}
           showfilters={filter}
-          showgroups={group}
+          showgroups={filter}
           showSelection={false}
         />{' '}
       </div>
@@ -362,7 +329,6 @@ const mapStateToProps = (state: IState): IStateProps => ({
 });
 
 const mapRecordsToProps = {
-  userPassages: (q: QueryBuilder) => q.findRecords('userpassage'),
   passages: (q: QueryBuilder) => q.findRecords('passage'),
   passageSections: (q: QueryBuilder) => q.findRecords('passagesection'),
   sections: (q: QueryBuilder) => q.findRecords('section'),
@@ -370,6 +336,6 @@ const mapRecordsToProps = {
   roles: (q: QueryBuilder) => q.findRecords('role'),
 };
 
-export default withData(mapRecordsToProps)(connect(mapStateToProps)(
-  TranscriptionTab
-) as any) as any;
+export default withData(mapRecordsToProps)(
+  connect(mapStateToProps)(TranscriptionTab) as any
+) as any;
