@@ -4,6 +4,7 @@ export * from './group';
 export * from './groupMembership';
 export * from './integration';
 export * from './invitation';
+export * from './mediaDescription';
 export * from './mediaFile';
 export * from './project';
 export * from './projectType';
@@ -14,8 +15,8 @@ export * from './passage';
 export * from './passageSection';
 export * from './user';
 export * from './role';
-export * from './userPassage';
 export * from './activityState';
+export * from './deleteItem';
 export * from './state';
 export * from '../store/book/types';
 export * from '../store/localization/types';
@@ -23,5 +24,6 @@ export * from '../store/localization/model';
 export * from '../store/orbit/types';
 export * from '../store/upload/types';
 export * from '../store/context/types';
-export * from '../store/langPicker/types';
 export * from '../store/media/types';
+export * from '../store/paratext/types';
+export * from '../store/auth/types';
