@@ -18,6 +18,7 @@ const schemaDefinition: SchemaSettings = {
         name: { type: 'string' },
         abbreviation: { type: 'string' },
         ownerId: { type: 'number' },
+        allUsers: { type: 'boolean' },
       },
       relationships: {
         owner: { type: 'hasOne', model: 'organization', inverse: 'groups' },
@@ -59,6 +60,9 @@ const schemaDefinition: SchemaSettings = {
       keys: { remoteId: {} },
       attributes: {
         email: { type: 'string' },
+        loginLink: { type: 'string' },
+        invitedBy: { type: 'string' },
+        strings: { type: 'string' },
         silId: { type: 'numer' },
         accepted: { type: 'boolean' },
         dateCreated: { type: 'date' },
@@ -71,6 +75,18 @@ const schemaDefinition: SchemaSettings = {
           model: 'organization',
         },
         role: {
+          type: 'hasOne',
+          model: 'role',
+        },
+        group: {
+          type: 'hasOne',
+          model: 'group',
+        },
+        groupRole: {
+          type: 'hasOne',
+          model: 'role',
+        },
+        allUsersRole: {
           type: 'hasOne',
           model: 'role',
         },
@@ -155,7 +171,7 @@ const schemaDefinition: SchemaSettings = {
           model: 'projecttype',
           inverse: 'projects',
         },
-        owner: { type: 'hasOne', model: 'user', inverse: 'projects' },
+        owner: { type: 'hasOne', model: 'user' },
         organization: { type: 'hasOne', model: 'organization' },
         group: { type: 'hasOne', model: 'group', inverse: 'projects' },
         projectIntegrations: {
@@ -163,7 +179,6 @@ const schemaDefinition: SchemaSettings = {
           model: 'projectintegration',
           inverse: 'project',
         },
-        users: { type: 'hasMany', model: 'userpassage', inverse: 'project' },
         // sections: { type: 'hasMany', model: 'section', inverse: 'project' },
         plans: { type: 'hasMany', model: 'plan', inverse: 'project' },
       },
@@ -232,7 +247,6 @@ const schemaDefinition: SchemaSettings = {
         sequencenum: { type: 'number' },
         book: { type: 'string' },
         reference: { type: 'string' },
-        position: { type: 'number' },
         state: { type: 'string' },
         hold: { type: 'boolean' },
         title: { type: 'string' },
@@ -244,11 +258,6 @@ const schemaDefinition: SchemaSettings = {
         sections: {
           type: 'hasMany',
           model: 'passagesection',
-          inverse: 'passage',
-        },
-        users: {
-          type: 'hasMany',
-          model: 'userpassage',
           inverse: 'passage',
         },
       },
@@ -277,6 +286,7 @@ const schemaDefinition: SchemaSettings = {
         filesize: { type: 'number' },
         dateCreated: { type: 'date' },
         dateUpdated: { type: 'date' },
+        position: { type: 'number' },
       },
       relationships: {
         plan: { type: 'hasOne', model: 'plan', inverse: 'mediafiles' },
@@ -289,26 +299,25 @@ const schemaDefinition: SchemaSettings = {
         name: { type: 'string' },
         givenName: { type: 'string' },
         familyName: { type: 'string' },
-        avatarUrl: { type: 'string' },
         email: { type: 'string' },
         phone: { type: 'string' },
         timezone: { type: 'string' },
         locale: { type: 'string' },
         isLocked: { type: 'boolean' },
-        externalId: { type: 'string' },
+        auth0Id: { type: 'string' },
+        silUserid: { type: 'number' },
         identityToken: { type: 'string' },
-        uiLanguageBcp47: { type: 'string' },
-        timerCountUp: { type: 'boolean' },
-        playBackSpeed: { type: 'number' },
-        progressBarTypeId: { type: 'number' },
+        uilanguagebcp47: { type: 'string' },
+        timercountUp: { type: 'boolean' },
+        playbackSpeed: { type: 'number' },
+        progressbarTypeid: { type: 'number' },
+        avatarUrl: { type: 'string' },
         hotKeys: { type: 'string' },
-        profileVisibilit: { type: 'number' },
-        emailNotification: { type: 'boolean' },
         dateCreated: { type: 'date' },
         dateUpdated: { type: 'date' },
+        lastModifiedBy: { type: 'number' },
       },
       relationships: {
-        projects: { type: 'hasMany', model: 'project', inverse: 'owner' },
         organizationMemberships: {
           type: 'hasMany',
           model: 'organizationmembership',
@@ -319,11 +328,6 @@ const schemaDefinition: SchemaSettings = {
           model: 'groupmembership',
           inverse: 'user',
         },
-        passages: {
-          type: 'hasMany',
-          model: 'userpassage',
-          inverse: 'users',
-        },
       },
     },
     currentuser: {
@@ -332,45 +336,30 @@ const schemaDefinition: SchemaSettings = {
         name: { type: 'string' },
         givenName: { type: 'string' },
         familyName: { type: 'string' },
-        avatarUrl: { type: 'string' },
         email: { type: 'string' },
         phone: { type: 'string' },
         timezone: { type: 'string' },
         locale: { type: 'string' },
         isLocked: { type: 'boolean' },
-        auth0: { type: 'string' },
-        // externalId: { type: 'string' },
-        // identityToken: { type: 'string' },
-        // uiLanguageBcp47: { type: 'string' },
-        // timerCountUp: { type: 'boolean' },
-        // playBackSpeed: { type: 'number' },
-        // progressBarTypeId: { type: 'number' },
-        // hotKeys: { type: 'string' },
-        // profileVisibilit: { type: 'number'},
-        // emailNotification: { type: 'boolean' },
+        auth0Id: { type: 'string' },
+        silUserid: { type: 'number' },
+        identityToken: { type: 'string' },
+        uilanguagebcp47: { type: 'string' },
+        timercountUp: { type: 'boolean' },
+        playbackSpeed: { type: 'number' },
+        progressbarTypeid: { type: 'number' },
+        avatarUrl: { type: 'string' },
+        hotKeys: { type: 'string' },
         dateCreated: { type: 'date' },
         dateUpdated: { type: 'date' },
+        lastModifiedBy: { type: 'number' },
       },
       relationships: {
-        projects: { type: 'hasMany', model: 'project' },
         organizationMemberships: {
           type: 'hasMany',
           model: 'organizationmembership',
         },
         groupMemberships: { type: 'hasMany', model: 'groupmembership' },
-      },
-    },
-    userpassage: {
-      keys: { remoteId: {} },
-      attributes: {
-        comment: { type: 'string' },
-        datecreated: { type: 'date' },
-        dateupdated: { type: 'date' },
-      },
-      relationships: {
-        role: { type: 'hasOne', model: 'role' },
-        user: { type: 'hasOne', model: 'user' },
-        passage: { type: 'hasOne', model: 'passage' },
       },
     },
   },
