@@ -4,8 +4,9 @@ import localizationReducer from './localization/reducers';
 import orbitReducer from './orbit/reducers';
 import uploadReducer from './upload/reducers';
 import contextReducer from './context/reducer';
-import langTagsReducer from './langPicker/reducers';
 import mediaReducer from './media/reducers';
+import paratextReducer from './paratext/reducers';
+import authReducer from './auth/reducers';
 
 import { composeWithDevTools } from 'redux-devtools-extension';
 // import { systemReducer } from "./system/reducers";
@@ -17,8 +18,9 @@ const appReducer = combineReducers({
   books: bookReducer,
   orbit: orbitReducer,
   upload: uploadReducer,
-  langTag: langTagsReducer,
   media: mediaReducer,
+  paratext: paratextReducer,
+  auth: authReducer,
 });
 
 export type AppState = ReturnType<typeof appReducer>;
@@ -38,5 +40,6 @@ export * from './book/actions';
 export * from './localization/actions';
 export * from './orbit/actions';
 export * from './upload/actions';
-export * from './langPicker/actions';
 export * from './media/actions';
+export * from './paratext/actions';
+export * from './auth/actions';
