@@ -34,7 +34,7 @@ const initialState = {
 			"invite": "Invite",
 			"action": "Action",
 			"delete": "Delete",
-			"filter": "Filter",
+			"filter": "Configure",
 			"selectRows": "Please select row(s) to {0}.",
 		}
 	}),
@@ -100,8 +100,10 @@ const initialState = {
 			"sections": "Sections",
 			"taks": "Passages",
 			"action": "Action",
+			"filter": "Configure",
 			"silTranscriberAdmin": "SIL Transcriber Admin",
 			"choosePlan": "Choose a Project Plan",
+			"loadingTable": "Loading data",
 		}
 	}),
 	"planSheet": new LocalizedStrings({
@@ -116,6 +118,9 @@ const initialState = {
 			"save": "Save",
 			"saving": "Saving",
 			"selectRows": "Please select row(s) to {0}.",
+			"confirm": "{0} {1} Item(s). Are you sure?",
+			"sectionAbove": "Insert section above",
+			"passageAbove": "Insert passage above",
 		}
 	}),
 	"scriptureTable": new LocalizedStrings({
@@ -126,6 +131,7 @@ const initialState = {
 			"book": "Book",
 			"reference": "Reference",
 			"description": "Description",
+			"loadingTable": "Loading data",
 		}
 	}),
 	"assignmentTable": new LocalizedStrings({
@@ -140,8 +146,7 @@ const initialState = {
 			"assignSec": "Assign Section",
 			"removeSec": "Remove Assignment",
 			"delete": "Delete",
-			"filter": "Filter",
-			"group": "Group",
+			"filter": "Configure",
 			"transcriber": "Transcriber",
 			"reviewer": "Reviewer",
 			"selectRowsToAssign": "Please select row(s) to assign.",
@@ -171,10 +176,10 @@ const initialState = {
 	"planAdd": new LocalizedStrings({
 		"en": {
 			"name": "Name",
-			"addPlan": "Add a Plan",
-			"newPlanTask": "Enter information for a new plan. (It could be a book of the Bible to be transcribed, a story, a lexionary, and so on.)",
+			"addPlan": "Add Plan",
+			"newPlanTask": "Type the name of the plan, which could be the name of a book of the Bible, a story, a lectionary and so on.",
 			"planType": "Type",
-			"selectPlanType": "Select plan type.",
+			"selectPlanType": "Choose the plan type.",
 			"cancel": "Cancel",
 			"add": "Add",
 			"save": "Save",
@@ -203,9 +208,11 @@ const initialState = {
 			"version": "Version",
 			"section": "Section",
 			"date": "Date",
-			"filter": "Filter",
+			"filter": "Configure",
 			"selectFiles": "Please select files to be uploaded.",
 			"selectRows": "Please select row(s) to {0}.",
+			"unsupported": "File {0} is an unsupported file type.",
+			"loadingTable": "Loading data",
 		}
 	}),
 	"passageMedia": new LocalizedStrings({
@@ -221,25 +228,43 @@ const initialState = {
 	"main": new LocalizedStrings({
 		"en": {
 			"silTranscriberAdmin": "SIL Transcriber Admin",
+			"silTranscriber": "SIL Transcriber",
 			"search": "Search…",
 			"organization": "Org Details",
 			"usersAndGroups": "Users and Groups",
 			"passages": "Passages",
 			"media": "Media",
+			"myTasks": "My Tasks",
+			"todo": "To Do",
+			"allTasks": "All Tasks",
+			"tasks": "Tasks",
 			"plans": "Plans",
 			"team": "Team",
 			"settings": "Settings",
 			"integrations": "Integrations",
 			"project": "Project",
-			"loadingTranscriber": "Loading SIL Transcriber Admin",
+			"loadingTranscriberAdmin": "Loading SIL Transcriber Admin",
+			"loadingTranscriber": "Loading SIL Transcriber",
 			"projectSummary": "Project Summary",
 			"addProject": "Add Project",
 			"logout": "Log Out",
 			"myAccount": "My Account",
 			"clearCache": "Clear cache",
+			"helpCenter": "Help Center",
+			"reportIssue": "Report an Issue",
+			"version": "Version: ",
 			"planUnsaved": "Plan Unsaved",
 			"loseData": "Do you want to leave this page and lose your changes?",
 			"newOrganization": "Add Organization",
+			"switchToAdmin": "Open project in Admin",
+			"switchToApp": "Open project in App",
+			"crashMessage": "Something went wrong. The developers need to address this issue.",
+			"reports": "Reports",
+			"sessionExpiring": "Session Expiring",
+			"sessionExpireTask": "Your session will expire in {0} seconds. Would you like to continue?",
+			"exit": "Exit",
+			"continue": "Continue",
+			"myWorkbench": "My Workbench",
 		}
 	}),
 	"transcriptionTab": new LocalizedStrings({
@@ -247,8 +272,7 @@ const initialState = {
 			"section": "Section",
 			"sectionstate": "State",
 			"passages": "Passages",
-			"filter": "Filter",
-			"group": "Group",
+			"filter": "Configure",
 			"transcriber": "Transcriber",
 			"reviewer": "Reviewer",
 		}
@@ -274,7 +298,7 @@ const initialState = {
 			"owner": "Owner",
 			"projects": "Projects",
 			"members": "Members",
-			"filter": "Filter",
+			"filter": "Configure",
 			"action": "Action",
 			"delete": "Delete",
 			"addGroup": "Add Group",
@@ -289,7 +313,7 @@ const initialState = {
 			"save": "Save",
 			"editGroup": "Edit Group",
 			"addGroup": "Add Group",
-			"newGroupTask": "Enter (or edit) basic group information.",
+			"newGroupTask": "Type the name and abbreviation of the group.",
 			"name": "Name",
 			"abbr": "Abbreviation",
 		}
@@ -300,18 +324,23 @@ const initialState = {
 			"abbreviation": "Abbreviation",
 			"save": "Save",
 			"projects": "Projects",
+			"owners": "Owners",
 			"reviewers": "Reviewers",
 			"transcribers": "Transcribers",
-			"addGroupMember": "Add Group Member",
-			"addMemberInstruction": "Enter the name of a person in the organization to be included in the group.",
+			"addGroupMember": "Choose {0}",
+			"addMemberInstruction": "Choose a person who will act as a {0} for projects in this group.",
 			"cancel": "Cancel",
-			"add": "Add",
+			"add": "Ok",
 			"delete": "Delete",
+			"choose": "Choose...",
 			"allReviewersCanTranscribe": "All Reviewers are allowed to transcribe.",
-			"noProjects": "There are no projects that use this group.",
+			"groupExplain": "Groups connect organization members to a project.",
+			"case1": "If the project does not exist, click the plus button in the Navigation pane to create it.",
+			"case2": "Existing projects can be associated with this group in Settings.",
 			"assignedSections": "  Assigned Sections: ",
 			"projectPlans": "Project Plans",
 			"invalidRole": "Invalid Role. User not added.",
+			"loadingTable": "Loading data",
 		}
 	}),
 	"shapingTable": new LocalizedStrings({
@@ -329,14 +358,14 @@ const initialState = {
 			"font": "Font",
 			"script": "Script",
 			"language": "Language",
-			"selectLanguage": "Select a Language",
+			"selectLanguage": "Choose Language Details",
 			"findALanguage": "Find a language by name, code, or country",
 			"codeExplained": "Code Explained",
 			"subtags": "Subtags",
 			"details": "Details",
 			"languageOf": "A Language of $1$2.",
 			"inScript": " in the $1 script",
-			"select": "Select",
+			"select": "Save",
 			"cancel": "Cancel",
 		}
 	}),
@@ -361,24 +390,40 @@ const initialState = {
 			"newInviteTask": "Enter the email address of the user to invite.",
 			"email": "Email",
 			"role": "Role",
-			"selectRole": "Select Organizational Role",
+			"organization": "Organization",
+			"selectOrgRole": "Select Organizational Role",
+			"allusersgroup": "All Users Group",
+			"groups": "Groups",
+			"group": "Group",
+			"additionalgroup": "Optional additional Group",
+			"groupRole": "Group Role",
+			"selectGroupRole": "Select Group Role",
 			"cancel": "Cancel",
-			"add": "Add",
+			"send": "Send",
 			"save": "Save",
 			"invalidEmail": "Invalid email address",
 			"alreadyInvited": "Already invited!",
+			"resend": "Resend",
+			"sil": "SIL International",
+			"transcriber": "SIL Transcriber",
+			"invitation": "has invited you to join",
+			"instructions": "Please click the following link to accept the invitation:",
+			"emailsubject": "SIL Transcriber Invitation",
 		}
 	}),
 	"invitationTable": new LocalizedStrings({
 		"en": {
 			"email": "Email",
 			"role": "Role",
+			"allUsers": "All Users",
+			"group": "Group",
 			"accepted": "Accepted",
 			"invite": "Invite",
 			"action": "Action",
 			"delete": "Delete",
-			"filter": "Filter",
+			"filter": "Configure",
 			"selectRows": "Please select row(s) to {0}.",
+			"noData": "No invitations",
 		}
 	}),
 	"orgSettings": new LocalizedStrings({
@@ -390,6 +435,135 @@ const initialState = {
 			"logo": "Logo",
 			"publicByDefault": "Public By Default",
 			"save": "Save",
+			"cancel": "Cancel",
+			"deleteOrg": "Delete Organization",
+			"deleteExplained": "Deleting this organization will prevent any members of the organization from doing work on any of the projects or plans in this organization.",
+		}
+	}),
+	"mediaUpload": new LocalizedStrings({
+		"en": {
+			"title": "Upload Media",
+			"task": "You can upload audio files in .mp3 format or in .wav format.",
+			"dragDrop": "Drag and drop files here, or click here to browse for the files.",
+			"cancel": "Cancel",
+			"upload": "Upload",
+		}
+	}),
+	"myTask": new LocalizedStrings({
+		"en": {
+			"todo": "To Do",
+			"history": "History",
+		}
+	}),
+	"toDoTable": new LocalizedStrings({
+		"en": {
+			"filter": "Configure",
+			"plan": "Plan",
+			"section": "Section",
+			"passage": "Passage",
+			"state": "State",
+			"action": "Action",
+			"transcriber": "Transcribe",
+			"reviewer": "Review",
+			"view": "View",
+			"length": "Duration",
+			"assigned": "Assigned",
+			"yes": "Yes",
+			"no": "No",
+			"loadingTable": "Loading data",
+			"title": "Title",
+			"description": "Description",
+			"duration": "Duration",
+		}
+	}),
+	"integration": new LocalizedStrings({
+		"en": {
+			"sync": "Sync",
+			"paratextAssociation": "Paratext association",
+			"selectProject": "Select Paratext Project",
+			"noProject": "You are not a member of any Paratext projects.",
+			"projectsPending": "Querying projects...",
+			"countPending": "Querying count...",
+			"usernamePending": "Querying username...",
+			"offline": "Offline",
+			"expiredToken": "Your login token has expired and can't be automatically renewed.  Please logout and login again.",
+			"expiredParatextToken": "Your paratext login has expired and can't be automatically renewed.  Please logout and login again.",
+			"questionOnline": "Are you connected to the Internet?",
+			"yes": "Yes",
+			"no": "No",
+			"questionProject": "Are you connected to a Paratext Project?",
+			"removeProject": "Remove Project Association",
+			"questionAccount": "Do you have a Paratext Account?",
+			"questionPermission": "Do you have permission to edit the Paratext project text?",
+			"allCriteria": "You must satisfy all criteria to sync.",
+			"projectError": "Project Query error:",
+			"usernameError": "Username error:",
+			"syncPending": "Syncing...",
+			"syncError": "Sync error:",
+			"syncComplete": "Sync Complete!",
+			"countError": "Count query error: ",
+			"countReady": "Passages ready to sync: ",
+			"paratext": "Paratext",
+			"render": "Render",
+			"onestory": "One Story",
+		}
+	}),
+	"transcriber": new LocalizedStrings({
+		"en": {
+			"submit": "Submit",
+			"reject": "Reject",
+			"save": "Save",
+			"close": "Close",
+			"settingsTip": "Settings",
+			"backTip": "Skip back (F2)",
+			"aheadTip": "Skip ahead (F4)",
+			"playTip": "Pause (ESC)",
+			"pauseTip": "Play (ESC)",
+			"rejectTranscriptionTip": "Audio recording is unusable",
+			"rejectReviewTip": "Transcription requires changes.",
+			"submitTranscriptionTip": "Mark this transcription as complete.",
+			"submitReviewTip": "Mark this review as complete.",
+			"saveTip": "Save transcription changes without marking as complete.",
+			"saveReviewTip": "Save position in recording without marking as complete.",
+		}
+	}),
+	"profile": new LocalizedStrings({
+		"en": {
+			"name": "Full Name",
+			"given": "Given Name",
+			"family": "Family Name",
+			"email": "Email",
+			"phone": "Phone",
+			"timezone": "Time zone",
+			"locale": "Locale",
+			"locked": "Locked",
+			"add": "Add",
+			"save": "Save",
+			"deleteUser": "Delete User",
+			"deleteExplained": "Deleting your user will block you from using the program and remove references to your work.",
+			"cancel": "Cancel",
+			"silTranscriberAdmin": "SIL Transcriber Admin",
+			"silTranscriber": "SIL Transcriber",
+			"userProfile": "User profile",
+			"completeProfile": "Complete User Profile",
+		}
+	}),
+	"deleteExpansion": new LocalizedStrings({
+		"en": {
+			"dangerZone": "Danger Zone",
+			"advanced": "Advanced",
+			"delete": "Delete",
+		}
+	}),
+	"taskItem": new LocalizedStrings({
+		"en": {
+			"noMedia": "No Audio",
+			"inProgress": "In Progress",
+			"transcribe": "Transcribe",
+			"review": "Review",
+			"sync": "Sync",
+			"done": "Done",
+			"section": "Section {0}.{1}",
 		}
 	}),
 };
@@ -430,6 +604,14 @@ export default function (state = initialState, action: any): ILocalizedStrings {
 				"invite" : new LocalizedStrings(action.payload.data.invite),
 				"invitationTable" : new LocalizedStrings(action.payload.data.invitationTable),
 				"orgSettings" : new LocalizedStrings(action.payload.data.orgSettings),
+				"mediaUpload" : new LocalizedStrings(action.payload.data.mediaUpload),
+				"myTask" : new LocalizedStrings(action.payload.data.myTask),
+				"toDoTable" : new LocalizedStrings(action.payload.data.toDoTable),
+				"integration" : new LocalizedStrings(action.payload.data.integration),
+				"transcriber" : new LocalizedStrings(action.payload.data.transcriber),
+				"profile" : new LocalizedStrings(action.payload.data.profile),
+				"deleteExpansion" : new LocalizedStrings(action.payload.data.deleteExpansion),
+				"taskItem" : new LocalizedStrings(action.payload.data.taskItem),
 			};
 		case SET_LANGUAGE:
 			return {
