@@ -1,4 +1,5 @@
-export const slug = (s: string) => {
+export const slug = (s: string | null) => {
+  if (!s) return '';
   return s
     .toLocaleLowerCase()
     .replace('&', '')
